+++ admin/src/scenes/missions/panel.jsx
@@ -1,16 +1,16 @@
-import React, { useState, useEffect } from "react";
-import styled from "styled-components";
-import { Link, useHistory } from "react-router-dom";
+import { React, useEffect, useState } from "react";
 import { toastr } from "react-redux-toastr";
+import { Link, useHistory } from "react-router-dom";
+import styled from "styled-components";
 
-import { translate, formatStringDateTimezoneUTC, MISSION_STATUS_COLORS, MISSION_STATUS } from "../../utils";
-import api from "../../services/api";
-import PanelActionButton from "../../components/buttons/PanelActionButton";
-import Panel, { Info, Details } from "../../components/Panel";
-import Badge from "../../components/Badge";
-import ModalConfirm from "../../components/modals/ModalConfirm";
-import { ROLES } from "snu-lib/roles";
 import { useSelector } from "react-redux";
+import { ROLES } from "snu-lib/roles";
+import Badge from "../../components/Badge";
+import Panel, { Details, Info } from "../../components/Panel";
+import PanelActionButton from "../../components/buttons/PanelActionButton";
+import ModalConfirm from "../../components/modals/ModalConfirm";
+import api from "../../services/api";
+import { MISSION_STATUS_COLORS, formatStringDateTimezoneUTC, translate } from "../../utils";
 
 export default function PanelView({ onChange, mission }) {
   const [tutor, setTutor] = useState();
@@ -65,14 +65,7 @@ export default function PanelView({ onChange, mission }) {
   };
 
   const onConfirmDuplicate = async () => {
-    mission.name += " (copie)";
-    delete mission._id;
-    mission.placesLeft = mission.placesTotal;
-    mission.status = MISSION_STATUS.DRAFT;
-    const { data, ok, code } = await api.post("/mission", mission);
-    if (!ok) toastr.error("Oups, une erreur est survnue lors de la duplication de la mission", translate(code));
-    toastr.success("Mission dupliquée !");
-    return history.push(`/mission/${data._id}`);
+    return history.push(`/mission/create/${mission.structureId}?duplicate=${mission._id}`);
   };
 
   if (!mission) return <div />;