+++ admin/src/scenes/team/list.jsx
@@ -177,6 +177,7 @@ export default function List() {
                     { label: "Date de création (récent > ancien)", field: "createdAt", order: "desc" },
                     { label: "Date de création (ancien > récent)", field: "createdAt", order: "asc" },
                   ]}
+                  selectedFilters={selectedFilters}
                   paramData={paramData}
                   setParamData={setParamData}
                 />