+++ admin/src/scenes/volontaires/list.jsx
@@ -129,6 +129,7 @@ export default function VolontaireList() {
                 { label: "Date de création (récent > ancien)", field: "createdAt", order: "desc" },
                 { label: "Date de création (ancien > récent)", field: "createdAt", order: "asc" },
               ]}
+              selectedFilters={selectedFilters}
               paramData={paramData}
               setParamData={setParamData}
             />