+++ api/src/controllers/mission.js
@@ -37,7 +37,7 @@ router.post("/", passport.authenticate("referent", { session: false, failWithErr
     const { error, value: checkedMission } = validateMission(req.body);
     if (error) {
       capture(error);
-      return res.status(400).send({ ok: false, code: ERRORS.INVALID_BODY });
+      return res.status(400).send({ ok: false, code: ERRORS.INVALID_BODY, error });
     }
 
     let structure = {};