+++ api/src/services/stats.service/engagement.js
@@ -494,6 +494,167 @@ async function getApplicationsChangeStatus(startDate, endDate, user) {
   }
 }
 
+async function getNewMissions(startDate, endDate, user) {
+  // ref dep only
+  let body = {
+    query: {
+      bool: {
+        filter: [
+          { terms: { "department.keyword": user.department } },
+          {
+            range: {
+              createdAt: {
+                gte: new Date(startDate),
+                lte: new Date(endDate),
+              },
+            },
+          },
+        ],
+      },
+    },
+    //to avoid duplicata
+    aggs: {
+      group_by_description: {
+        terms: { field: "description.keyword", size: ES_NO_LIMIT },
+      },
+    },
+    size: 0,
+    track_total_hits: true,
+  };
+
+  const response = await esClient.search({ index: "mission", body });
+  const value = response.body.aggregations.group_by_description.buckets.length;
+
+  return [
+    {
+      id: "new-missions",
+      value: value,
+      label: ` nouvelle${value > 1 ? "s" : ""} mission${value > 1 ? "s" : ""} déposée${value > 1 ? "s" : ""}`,
+      icon: "action",
+    },
+  ];
+}
+
+async function getProposedMissionsAcceptedOrRefusedByYoung(startDate, endDate, user) {
+  // ref dep only
+  const token = await getAccessToken(API_ANALYTICS_ENDPOINT, API_ANALYTICS_API_KEY);
+  let body = {
+    startDate: formatDateForPostGre(startDate),
+    endDate: formatDateForPostGre(endDate),
+    status: [APPLICATION_STATUS.VALIDATED, APPLICATION_STATUS.REFUSED],
+    department: user.department,
+  };
+
+  const response = await fetch(`${API_ANALYTICS_ENDPOINT}/stats/application-accepted-refused/count`, {
+    ...postParams(token),
+    body: JSON.stringify(body),
+  });
+
+  const result = await response.json();
+  const data = result?.data;
+  let resultArray = {
+    [APPLICATION_STATUS.VALIDATED]: 0,
+    [APPLICATION_STATUS.REFUSED]: 0,
+  };
+
+  for (const item of data) {
+    const value = item.value;
+    if (Object.prototype.hasOwnProperty.call(resultArray, value)) {
+      resultArray[value]++;
+    }
+  }
+  const validatedValue = resultArray[APPLICATION_STATUS.VALIDATED];
+  const refusedValue = resultArray[APPLICATION_STATUS.REFUSED];
+
+  return [
+    {
+      id: "applications-validated",
+      value: validatedValue,
+      label: ` proposition${validatedValue > 1 ? "s" : ""} de missions acceptée${validatedValue > 1 ? "s" : ""} par des volontaires`,
+      icon: "action",
+    },
+    {
+      id: "applications-refused",
+      value: refusedValue,
+      label: ` proposition${refusedValue > 1 ? "s" : ""} de missions refusée${refusedValue > 1 ? "s" : ""} par des volontaires`,
+      icon: "action",
+    },
+  ];
+}
+
+async function getYoungStartPhase2InTime(startDate, endDate, user) {
+  // ref reg only
+  let body = {
+    query: {
+      bool: {
+        filter: [
+          { term: { "region.keyword": user.region } },
+          { terms: { "statusPhase2.keyword": [YOUNG_STATUS_PHASE2.VALIDATED, YOUNG_STATUS_PHASE2.IN_PROGRESS] } },
+          { terms: { "statusPhase2Contract.keyword": [CONTRACT_STATUS.VALIDATED] } },
+        ],
+      },
+    },
+    size: ES_NO_LIMIT,
+    track_total_hits: true,
+  };
+
+  const response = await esClient.search({ index: "young", body });
+  const youngs = response.body.hits.hits;
+  const youngsIds = youngs.map((young) => young._id);
+  const youngsCohort = youngs.map((young) => ({
+    _id: young._id,
+    cohort: young._source.cohort,
+  }));
+
+  let body2 = {
+    query: {
+      bool: {
+        filter: [{ terms: { "youngId.keyword": youngsIds } }, { term: { "youngContractStatus.keyword": CONTRACT_STATUS.VALIDATED } }],
+      },
+    },
+    aggs: {
+      group_by_youngId: {
+        terms: { field: "youngId.keyword", size: ES_NO_LIMIT },
+        aggs: {
+          minYoungContractValidationDate: {
+            min: { field: "youngContractValidationDate" },
+          },
+        },
+      },
+    },
+    size: 0,
+    track_total_hits: true,
+  };
+
+  const response2 = await esClient.search({ index: "contract", body: body2 });
+
+  let value = response2.body.aggregations.group_by_youngId.buckets;
+
+  value = value.filter((bucket) => {
+    const minYoungContractValidationDate = new Date(bucket.minYoungContractValidationDate.value);
+    const correspondingYoung = youngsCohort.find((young) => young._id === bucket.key);
+
+    //check si la date de validation de contract est moins d'un an après la date de validation de phase 1 du jeune
+    if (correspondingYoung) {
+      const oneYearInMilliseconds = 365 * 24 * 60 * 60 * 1000;
+      const cohortEnd = COHESION_STAY_END[correspondingYoung.cohort];
+
+      if (minYoungContractValidationDate - cohortEnd < oneYearInMilliseconds) {
+        return minYoungContractValidationDate >= new Date(startDate) && minYoungContractValidationDate <= new Date(endDate);
+      }
+    }
+  });
+
+  return [
+    {
+      id: "young-phase2-validated-in-time",
+      value: value.length,
+      label: ` volontaire${value.length > 1 ? "s" : ""} ayant commencé ${value.length === 1 ? "sa" : "leur"} mission d’intérêt général dans le délai imparti`,
+      icon: "action",
+    },
+  ];
+}
+
 module.exports = {
   getYoungNotesPhase2,
   getNewStructures,
@@ -503,4 +664,7 @@ module.exports = {
   getYoungsWhoStartedOrFinishedMissions,
   getMissionsChangeStatus,
   getApplicationsChangeStatus,
+  getNewMissions,
+  getProposedMissionsAcceptedOrRefusedByYoung,
+  getYoungStartPhase2InTime,
 };