+++ api/src/services/stats.service/index.js
@@ -8,6 +8,9 @@ const {
   getYoungsWhoStartedOrFinishedMissions,
   getMissionsChangeStatus,
   getApplicationsChangeStatus,
+  getNewMissions,
+  getProposedMissionsAcceptedOrRefusedByYoung,
+  getYoungStartPhase2InTime,
 } = require("./engagement");
 const { getYoungNotesPhase1, getTimeScheduleAndPedagoProject, getTransportCorrectionRequests, getSessions, getLineToPoints } = require("./sejour");
 const {
@@ -61,8 +64,10 @@ const keyNumbersByRole = {
       getYoungsWhoStartedOrFinishedMissions,
       getMissionsChangeStatus,
       getApplicationsChangeStatus,
+      getNewMissions,
+      getProposedMissionsAcceptedOrRefusedByYoung,
     ],
-    [ROLES.REFERENT_REGION]: [getYoungPhase2Validated],
+    [ROLES.REFERENT_REGION]: [getYoungPhase2Validated, getYoungStartPhase2InTime],
     [ROLES.SUPERVISOR]: [getMissionsOnTerm, getYoungsWhoStartedOrFinishedMissions, getMissionsChangeStatus, getApplicationsChangeStatus],
     [ROLES.RESPONSIBLE]: [getMissionsOnTerm, getYoungsWhoStartedOrFinishedMissions, getMissionsChangeStatus, getApplicationsChangeStatus],
     [ROLES.ADMIN]: [getContractsSigned, getYoungsWhoStartedOrFinishedMissions, getMissionsChangeStatus, getApplicationsChangeStatus],
@@ -82,6 +87,8 @@ const keyNumbersByRole = {
       getYoungValidatedFromWaitingStatus,
       getYoungWhoChangedCohort,
       getApplicationsChangeStatus,
+      getNewMissions,
+      getProposedMissionsAcceptedOrRefusedByYoung,
     ],
     [ROLES.REFERENT_REGION]: [
       getTimeScheduleAndPedagoProject,
@@ -93,6 +100,7 @@ const keyNumbersByRole = {
       getAbandonedRegistration,
       getYoungValidatedFromWaitingStatus,
       getYoungWhoChangedCohort,
+      getYoungStartPhase2InTime,
     ],
     [ROLES.SUPERVISOR]: [getMissionsOnTerm, getYoungsWhoStartedOrFinishedMissions, getMissionsChangeStatus, getApplicationsChangeStatus],
     [ROLES.RESPONSIBLE]: [getMissionsOnTerm, getYoungsWhoStartedOrFinishedMissions, getMissionsChangeStatus, getApplicationsChangeStatus],