+++ api/src/services/stats.service/inscription.js
@@ -282,6 +282,7 @@ async function getYoungValidatedFromWaitingStatus(startDate, endDate, user) {
 
   const result = await response.json();
   const data = result?.data;
+  console.log(`🐞 - file: inscription.js:285 - data:`, data);
   let resultArray = {
     [YOUNG_STATUS.WAITING_VALIDATION]: 0,
     [YOUNG_STATUS.WAITING_LIST]: 0,