+++ app/src/context/RepresentantsLegauxContextProvider.jsx
@@ -1,5 +1,5 @@
 import React, { createContext, useEffect, useState } from "react";
-import { useHistory } from "react-router-dom";
+import { Redirect } from "react-router-dom";
 import queryString from "query-string";
 import api from "../services/api";
 import { capture } from "../sentry";
@@ -7,28 +7,27 @@ import { capture } from "../sentry";
 export const RepresentantsLegauxContext = createContext();
 
 const RepresentantsLegauxContextProvider = ({ children, parentId }) => {
-  const history = useHistory();
+  const [error, setError] = useState(false);
   const params = queryString.parse(location.search);
   const { token } = params;
   const [young, setYoung] = useState(null);
 
   useEffect(() => {
     async function getYoungFromToken() {
-      const redirectInvalidToken = () => history.push("/representants-legaux/token-invalide");
       try {
-        if (!token) redirectInvalidToken();
+        if (!token || !parentId) setError(true);
         const { ok, data } = await api.get(`/representants-legaux/young?token=${token}&parent=${parentId}`);
-        console.log("ok", ok, data);
-        if (!ok) return redirectInvalidToken();
+        if (!ok) return setError(true);
         setYoung(data);
       } catch (e) {
-        if (e.code === "INVALID_BODY" || e.code === "OPERATION_UNAUTHORIZED") return redirectInvalidToken();
+        if (e.code === "INVALID_BODY" || e.code === "OPERATION_UNAUTHORIZED") setError(true);
         capture(e);
       }
     }
     getYoungFromToken();
-  }, []);
+  }, [token, parentId]);
 
+  if (error) return <Redirect to="/representants-legaux/token-invalide" />;
   return <RepresentantsLegauxContext.Provider value={{ young, token }}>{children}</RepresentantsLegauxContext.Provider>;
 };