+++ app/src/scenes/inscription2023/desktop/stepCoordonnees.jsx
@@ -322,7 +322,7 @@ export default function StepCoordonnees() {
       } catch (error) {
         capture(error);
       }
-    }, 500),
+    }, 100),
     [],
   );